A command to change chat message /ChatColor HEX
I finally figured out a way to do so! :P might be usefull for users who
want to select colors of their choice for chat messages!
/ChatColor HEX
It reformat the format string for that i have a _reformat hook for
format and SendClientMessage/ToAll. Working of replacement done through
format_fix_color.